requirements.txt
python-dotenv
requests
pandas
tradingview-ta
numpy
flask
psutil
